requirements.txt
flask
python-dotenv
pandas
groq
